fix: more compiler warnings

+++ include/tev/Common.h
@@ -341,7 +341,7 @@ inline float toSRGB(float linear, float gamma = 2.4f) {
     if (linear <= 0.0031308f) {
         return 12.92f * linear;
     } else {
-        return (1 + a) * pow(linear, 1 / gamma) - a;
+        return (1.0f + a) * (float)pow(linear, 1.0f / gamma) - a;
     }
 }
 
@@ -350,7 +350,7 @@ inline float toLinear(float sRGB, float gamma = 2.4f) {
     if (sRGB <= 0.04045f) {
         return sRGB / 12.92f;
     } else {
-        return pow((sRGB + a) / (1 + a), gamma);
+        return (float)pow((sRGB + a) / (1.0f + a), gamma);
     }
 }
